pyproject.toml
[build-system]
requires = ["uv_build>=0.9.4,<0.10.0"]
build-backend = "uv_build"

[project]
name = "pydantic-temporal-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pydantic-ai-slim[duckduckgo]>=1.0.2",
    "logfire[fastapi]>=4.3.4",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "slack-sdk>=3.36.0",
    "typer>=0.12.3",
    "pydantic-settings>=2.4.0",
    "httpx>=0.27",
    "uvloop>=0.20",
    "pygithub>=2.8.1",
    "pydantic-ai-claude-code>=0.8.1",
    "aiohttp>=3.13.2",
]

[project.scripts]
temporal-agent = "src.pydantic_temporal_example.cli:app"

[dependency-groups]
dev = [
    # Needed for PyCharm users
    "pip>=25.2",
]
lint = ["pyright>=1.1.405", "ruff>=0.12.10"]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
]

[tool.uv]
default-groups = ["dev", "lint", "test"]

[tool.ruff]
line-length = 120
target-version = "py313"
include = [
    "src/**/*.py",
]

[tool.ruff.lint]
extend-select = [
    "Q",
    "RUF100",
    "RUF018", # https://docs.astral.sh/ruff/rules/assignment-in-assert/
    "C90",
    "UP",
    "I",
    "D",
    "TID251",
]
mccabe = { max-complexity = 15 }
ignore = [
    "D100", # ignore missing docstring in module
    "D102", # ignore missing docstring in public method
    "D104", # ignore missing docstring in public package
    "D105", # ignore missing docstring in magic methods
    "D107", # ignore missing docstring in __init__ methods
    "S104", # ignore binding to all interfaces (needed for containers)
    "TRY300", # ignore moving return to else block (conflicts with RET505)
]

[tool.ruff.lint.isort]
combine-as-imports = true
known-first-party = ["src"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pyright]
pythonVersion = "3.14"
typeCheckingMode = "strict"
reportMissingTypeStubs = false
reportUnnecessaryIsInstance = false
reportUnnecessaryTypeIgnoreComment = true
reportMissingModuleSource = false
include = [
    "src",
]
venvPath = '.'
venv = ".venv"
